FEAT: Make canvas to svg and override each other
